Fixed some of the project structure
Physico's GL object is now a JSON standalone object, not a function object. Love the singleton things ;))

Plus, now the shaders are loaded dynamically (XHttpRequest) from static files. This is going to get useful when we start working more and more on the GL side, and further from Javascript.

I think we're ready for a new tag version.

Signed-off-by: Sabin Marcu <sabinmarcu@gmail.com>